Move the cleanup methods in util to netlink pkg
